refactor: 에러 메시지를 상수 처리로 이동 (constant 패키지 수정)
