+++ src/main/java/calculator/constant/OutputMessage.java
@@ -1,13 +1,13 @@
 package calculator.constant;
 
-public enum MessageConstant {
+public enum OutputMessage {
 
     INPUT_ADDITION("덧셈할 문자열을 입력해 주세요. "),
     ADDITION_RESULT("결과 : %d");
 
     private final String message;
 
-    MessageConstant(String message) {
+    OutputMessage(String message) {
         this.message = message;
     }
