+++ src/main/java/calculator/logic/CalculatorComputeResult.java
@@ -1,4 +1,14 @@
 package calculator.logic;
 
+import java.util.List;
+
 public class CalculatorComputeResult {
+    public Integer computeResult(List<Integer> integerList) {
+        Integer result = 0;
+        for (Integer num : integerList) {
+            result += num;
+        }
+        return result;
+    }
+
 }
